backend/requirements.txt
# Agentic MCP Gateway — Dependencies
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
pydantic>=2.10.4
python-dotenv>=1.0.1
groq>=0.25.0
sse-starlette>=2.2.1
httpx>=0.28.1
slack_sdk>=3.0.0
gspread>=6.0.0
google-auth>=2.0.0
